Remove empty 'gi' handler; make code work for different prefixes.
The prefix is now set based on the `event.which` number, and is
used in a conditional in the `handlePrefixShortcuts` method to
determine which shortcuts are called by that particular prefix.

This would allow for easily handling other prefix keys, by adding
more conditions or a case statement in the `handlePrefixShortcuts`
method.

Also removed the empty handler for the 'gi' shortcut, since that
isn't being used yet.