Refactor the initial setting of focus and set it with 'j' and 'gg'.
The `moveFocus` method is now doing one of four types of movements,
so it shouldn't also be tasked with placing the initial focus if
there is no `focused` element yet.

This commit adds a helper method and calls it if focus is not placed yet
for either the 'j' or 'gg' shortcuts.

Issues:
 * Before each movement shortcut, we want to set the focus if it hasn't
 already been set.  The helper method to do this can be called as the
 optional first step inside of `moveFocus` rather than calling it every
 time that `moveFocus` fails.