Refactors the setting of focus when no focus is set.
When using keyboard shortcuts, the focus should be set the first time
a movement keyboard shortcut is used.  For example, 'G' should both
scroll to the bottom and set the focus on the last element if no focus
had been set already.

Since checking and setting the focus is a prerequisite for any movement,
this commit adds a helper method to get/set the focus and uses the
helper method at the start of the `moveFocus` method.