Fixes issue with j shortcut jumping to second focusable item.
Previous changes refactored the moving of focus so that each time a
movement shortcut is hit, the `moveFocus` method sets the focus if it
wasn't already set.  It then makes the appropriate movement called for
by the shortcut.  The meant that on a fresh page, the 'j' shortcut
caused the focus to be first set at the top, and then move forward by
one.  The desired behavior would be for 'j' to focus the first item
on a fresh page.

This commit adds a check for this special case - if the focus was not
set when `moveFocus` was called, and the movement type is `forward`,
then it stops after setting the focus on the next item.

After cleaning this up, a next step could be further refactoring the
handling of focus.  It needs to be further simplified, and this is a
temporary fix.