Revert "Last sweep" box, add column to projects table
This reverts most of ced09cd15cf8691a046b31401a07f182ee6dce45. I gave it
some runtime and didn't like how it made the page. @lencioni suggested
that we put a "Last sweep" column in the projects table instead, and I
agree about that being a better solution. This commit adds that column
and reverts the "Last sweep" box.

There's an N+1 introduced here (each project in the table needs to fetch
it's latest sweep from the db). It's probably fine, but I don't like how
easy it is to introduce perf regression over time. I intend to follow up
with a commit that starts caching the last sweep id on the projects
table, and therefor making it possible to avoid the N+1 with a single
`includes` call when fetching projects.