Add `last_sweep` column cache for projects
This is an optimization to avoid making N+1 queries when rendering the
projects index page. It does so by keeping a column cache on the
project referencing the last created sweep.

To backfill this column, you can run this in a shell:

    bundle exec rails runner 'Project.all.each { |p| p.refresh_last_sweep! }' }'

If you don't backfill, you just won't see any "Last sweep" progress
bars. Nothing will break.