Update Rails 4.0.4 -> 4.1.0
This upgrade was very straightforward. I followed the advice in the
upgrade guide[1] to remove the line from the spec helper since it will
automatically happen as of 4.1.0.

The one section that looked interesting but will require further
investigation before we implement it here is the addition of the
secrets.yml file.

[1]: http://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#upgrading-from-rails-4-0-to-rails-4-1

Change-Id: I2e16844385be0a707215a20e7a00933e35742c57