Add helper function to turn off transactions for some specs
When using poltergeist to run feature specs with PhantomJS as the
driver, the default RSpec transactions don't work.  In order to
create test data in the database to prepare for specs, and to clear
that test data after the specs run, a helper function sets up and
then removes custom configurations for handling the test database.

The helper method relies on the 'DatabaseCleaner' gem for manually
handling the timing and method of cleaning the test data out of the
test database.

This should have no effect on existing specs where the helper method
is not called.