Add feature specs for movement keyboard shortcuts (j, k, gg, G)
The keyboard shortcuts 'j', 'k', 'gg', and 'G' all either place and/or
move the focus on a page with multiple focusable elements.
This spec adds testing of those keyboard shortcuts in the context of
two different pages on the site, running with the headless browser
PhantomJS.  This is part of implementing feature specs for
keyboard shortcuts in general, to make future refactoring of the
related JavaScript easy to do without breaking the application.

Potential Issues:
 * The feature specs run a bit slower than the other specs because they
 use custom configuration which cleans the database with truncation.

Next Steps:
 * Write specs for other keyboard shortcuts.
 * Refactor the JavaScript fearlessly!