Fix diff toolbar gradient in Firefox
We had originally implemented this as a gradient that went from white to
"transparent". While this looked good in Chrome, in Firefox it looked
oddly grayish. I believe that Firefox was transitioning to transparent,
but since it didn't know what color of transparent, it transitioned to
mid-gray at full transparency, causing a strange gray color.

To fix this, I changed the color that we transition to in the gradient
from "transparent" to rgba(255, 255, 255, 0), which is fully-transparent
white.

Closes #100.

Change-Id: I4e40a0c8671dea207916d940ebf8b51b9115cb39