Allow elements to specify multiple keyboard shortcuts
We want to add `J` and `K` as aliases for `]` and `[`, respectively. To
do this, we first need to allow these elements to have multiple keyboard
shortcuts.

The `~=` in the attribute selector will match elements whose attribute
value is a list of space-separated values, one of which is exactly equal
to the value given in the selector. This will enable the functionality
we need for this feature.

Addresses #99.

Change-Id: Iaa7e5b9c6df755fcbef6c68b9f6088801fb77fee