Dynamically associate shortcut keycode with character
We had an object that mapped keycodes that we have been using as
shortcut keys to their character strings. However, we can do this
dynamically via `String.fromCharCode()`. Although there will be a
performance penalty for this, I don't think avoiding that penalty
warrants the extra complexity at this time.

In service of #99.

Change-Id: Ib1cae0ab3257da4406953df53c50af8acd0488de