Re-compare with same snapshot as before
When you hit the "Re-compare" button on the snapshot page, we used to
compare the snapshot against the current baseline and not the baseline
that was at the time of first comparison. This was confusing, and led to
situations where hitting that button would remove the old diff and
either replace it with a diff to some other snapshot, or with no new
diff at all, since we only compare snapshots with a baseline if the
baseline is older than the snapshot.

I've fixed this by caching the compared_with snapshot and reusing that
when re-comparing. I'm aware of the fact that the snapshot_diff model
already stores this information, but we wipe out that reference when
re-comparing, and I would have had to make larger structural changes in
order to use that information.