Fix broken snapshots controller spec
I introduced a bug in my last commit by changing the characteristics of
the snapshot under test. I did that to make sure that the buggy behavior
of triggering a comparer when retaking the snapshot was taken care of.
By using a combination of :with_baseline and :with_diff traits I fixed
the spec and at the same time kept the validation that the comparer
wasn't triggering.