Animate snapshot thumbnail while still processing
Before a snapshot has been taken, we render a clock icon and animate its
opacity up and down to signify that it is actively being worked on and
is not ready to be reviewed. We should keep animating these after the
snapshot has been taken, before the diff has been generated to clearly
communicate that these items are not ready for review yet.

Closes #101.

Change-Id: I5118b9a87d0d7ae03b56383b7298400fe87bfa78