Add patch to allow testing `click` in JS with phantomJS
PhantomJS does not support a native `click` method for HTMLElement.
Adding a patched version of this allows the JS feature specs to
execute `click` without overwriting the `click` method that is
available when running the code in modern browsers.

Without this patch, a call to `click` in the JavaScript will cause
specs to fail with a `undefined is not a method` error.