Add snapshot animations
This commit adds a controller that cycles through the accepted images
with > 0% diffs and put together an animation using ImageMagick. This
can give you an interesting look at how a page has evolved over time.

This is a rough swing at this and is by no means perfect. The biggest
problem it currently has is dealing with transparency, either within the
snapshot, or when the snapshot changes height.

We could also likely do something smarter about how we are storing these
images, perhaps with a model and paperclip, but I wanted to just get
this running with minimal fuss.

It is also worth noting that these images can end up being quite large.
It would probably be worth looking into methods for producing smaller
animations.

Closes #29.

Change-Id: I40d90b920ac16f511a236250476b0d6e1d5b62e8