Escape html in sweep descriptions
We noticed that the word "<b>" didn't show up for the description in a
sweep we ran at causes. Instead everything after it was bolded. We
decided to escape the text instead. We were already passing the text
through `simple_format`, and that helper method does some sanitization.
But it won't get rid of some of the most common html tags.

While changing this, I noticed that we wrapped the `simple_format` call
with a <p class="help"> tag. That led to invalid markup (a p tag wrapped
by a p) so I moved the `help` class into the call to `simple_format`.