Store a log from the snapshotting process
To better be able to debug snapshots that look wrong, I'm making the log
output from take-snapshot.js bubble up and be persisted in the snapshot
model. To access the log, hit /snapshots/{id}/view_log in your browser
or use the "View debug log" in the footer of the snapshot page (it's
only there if there is a log).

I hope this helps us understand if we need to up the timeouts defined in
the take-snapshot.js script.