Increase fallback timeout to 60 seconds
After I started saving a debug log for each snapshot, I found that when
you're running a sweep, a couple of pages will take more than 20 seconds
to load and snapshot. The fallback timeout is there to make sure that
the snapshot is taken even if something goes wrong. In some cases
though, the page could still be downloading resources when this timeout
occurs.

60 seconds seems like a better fit for now. It's probably too much but
to be honest I don't expect this to fire very often so a long timeout is
probably fine.