Fix trivial JS error with sweep pager buttons
In Safari (and by extension in PhantomJS), clicking a link that
has `javascript return false;` triggers an error.

This error had no effect on functionality, but was causing the new
feature specs for the "[" and "]" shortcuts to fail.  Changing the
links to simply return nothing allows the same functionality with
no javascript errors.