Add specs for "[" and "]"
In order to write specs for shortcuts dealing with snapshots and
sweeps, a new trait was added to the sweep factory to generate
a sweep with a full set of associated snapshots.

These specs revealed a minor JS error that was not breaking
functionality, and is fixed in a separate commit.