Add final keyboard shortcut specs (for 'a', 'r', and 'x')
The final section of keyboard shortcut specs covers the shortcuts
for navigating, accepting, and rejecting snapshot diffs.

In order to have the links for `Before` and `After` views of a
snapshot diff in the test data, an `image_width` property was added
to the `snapshot_diff` association on the `snapshot` factory trait
of `with_diff`.

In other words, when the snapshot factory is used with the `with_diff`
trait, it will now generate snapshot diffs with images that have an
`image_width` property and return `true` for `sprite?`.

Branch Summary:
This is the final commit in a branch that adds specs for all the
current keyboard shortcuts.
In order to add these feature specs, the early commits add `Poltergeist`
and some custom configurations for RSpec and Capybara.  The specs run
with PhantomJS, which emulates using something like Safari to click
through the site.