Update README to not reference trotzig/diffux
Now that we've moved into an organization profile, we want documentation
to point to the right place. I've updated links and 3rd-party
integrations so that they're not using the old repo home.

Note: There is one reference to trotzig/diffux remaining -- the rubydoc
link. I'm not entirely sure how to set that up. @lencioni set it up in
the first place so I think I'll just consult him.