Update documentation URL
Now that we have moved the repo to an organization, we want all of our
links to be pointing to the correct place. 00e98abac0bbcc32e7135b
updated many of the links, but not the documentation one. This commit
does that.

Additionally, I decided to remove the frames option from the link
because nobody likes frames.

Change-Id: I1010d8392d036393c2cf1bb512770dac4810cc69