Detach Snapshot model from SnapshotComparer
I'm in the process of extracting a diffux-core repository with tools
that help diffing and snapshotting, but doesn't have a dependency to the
entire Diffux application.

The SnapshotComparer service didn't need a lot of changes to make this
happen. I made it so that you have to pass in ChunkyPNG images instead
of Snapshot objects. The specs got simpler (and faster!), and I also
DRY:ed them up a but a bit by creating an `image` helper method.