i18nalize snapshots/show
All of the copy for this section has now been extracted to the en.yml
configuration file.

I added `Snapshot#before_snapshot` method because of the Law of Demeter
and to reduce the amount of code in the view. There are probably a few
other opportunities like this in here that we should loop back and take
care of.

Change-Id: If89476ee7112e12dd6b73a7e32dd57eb16cc7f6f