Start depending on `diffux-core`
I've copied the core of Diffux into a new repository [1]. This commit
will clean out the short-lived code duplication and add diffux-core as a
dependency. For now, I'm depending on the latest from the github repo
itself. When I've tested out some more I'll publish a 0.0.1 version of
diffux-core and start depending on that instead.

[1]: https://github.com/diffux/diffux-core