i18nalize shared/error_explanation
The copy for this partial is now in our en.yml configuration file. This
will also render the model validation messages which I did not add to
the en.yml configuration file because they are auto-generated by Rails
in English. However, when we add translations, we probably need to
translate these messages in each language's config file. According to
the Rails i18n guide[1], the key that Rails looks for follows this
pattern:

  activerecord.errors.models.[model_name].attributes.[attribute_name]

[1]: http://guides.rubyonrails.org/i18n.html

Change-Id: Id1151ad516afeb3afacf4b9a503f3a7b38340d61