Set up URL-based locale switching
This makes switching the locale of the app as easy as changing the first
part of the URL. By default it puts you in en. We currently have a
Spanish translation which you can switch to by changing the "en" in the
URL to "es".

We will probably want to at some point add an easy way in the app to
jump to a different locale, but in the meantime this seems pragmatic.
Also, instead of defaulting to en, it might be nice to extract the
locale from the HTTP_ACCEPT_LANGUAGE header and redirect there to begin
with.

Change-Id: I9432f3da7cfbe4cf95248aab4c57b11d7b793dc0