Un-scope refresh route
I recently scoped all routes under the locale. Well, the refresh route
is referenced by JavaScript to auto update snapshots and progress bars
that are waiting to be finished. This route always returns a JSON
response that includes no localized copy, so we can just un-scope it to
make it work again.

Change-Id: If446e67da0013223f0139a9982272289b993c6c8