Allow locale in URL to be optional
When I first added URL-based locale specification, I made it required
and had `/` redirect to `/en`. However, I realized that it would be
better to just make it optional. This will allow routes like refresh and
sweep triggering to work as they did without any change, but to also
optionally specify a locale.

Change-Id: Ib75ffe8bb4c2396dd41f9e89d788ed368255682c