Fix sweep progress bars when using es locale
We're using the `number_to_percentage` helper to construct the
css/style string used to control the width of the individual bars inside
the progress bar. This method is sensitive to locale, so when in "es"
locale, the delimiter would be a comma instead of a dot. This caused
invalid css. I've fixed that by passing in an "en" locale to the helper.