Fix keyboard shortcuts spec
When adding the locale scope to the routes, all of the path and url
helpers that passed in an argument needed to be adjusted so that Rails
understands that the argument is the id and not the locale param.
Unfortunately, I failed to do this in the keyboard shortcuts feature
spec, so it has been broken for a bit. This commit resolves this
problem.

Change-Id: Ia9ae1673189870204f79ae44b1f20c20553c84b9