Set default locale in routes
After adding the optional locale param scope to the routes, a bunch of
controller specs blew up. It looked like RSpec thought that the ID param
in GET requests was the locale param, causing errors along the lines of
"No route matches...".

After some digging around, I stumbled upon an issue[1] where someone was
having a similar problem. @noplay suggested a workaround that involved
setting the default in the routes when running tests. I played around
with a few options and landed on a solution where I specify the default
locale param value as I18n.default_locale. This fixes the broken tests
and allows us to run with the same configuration in all environments
without any problems.

[1]: https://github.com/rspec/rspec-rails/issues/562

Change-Id: I9e340810f1af4832cbdc42935ca4e6a71677266e