Fix bug with progress bar when sweep is delayed
When a delayed sweep started, the progress bar showed a broken HTML
string 'unknown">' due to a lack of translation string for the title of
a sweep with this status.  The class names for the progress bar layers
were also coming in with 'NaN', due to a divide-by-0 error.

This commit fixes both of those issues by adding a translation string
for the status of a delayed sweep ('Delayed') and checking that the
total number of sweeps is greater than 0 before dividing the
sub-totals.