Update Heroku documentation
It's fallen a bit out of date due to the need to set the BUILDPACK_URL
environment variable.

Other than that, I rephrased the instructions a bit and gave better
information about how to not spend lots of money if your keys are
compromised, and how to not spend little amounts of money if you forget
to scale down the worker process when you're not using it.