Update Rails 4.1.1 -> 4.2.1
There seems to have been a bug in Rails 4.1.0 that was fixed in 4.1.2
that was affecting fresh installs of diffux with an error like:

> NoMethodError: undefined method `name' for nil:NilClass

This was related to the counter cache, and upgrading to the latest
version of Rails resolves it.

I first upgraded Rails without upgrading turbolinks, but I ran into
problems with `redirect_to` taking the wrong number of arguments and
upgrading turbolinks seemed to resolve this.

This upgrade seems to come with a lot of deprecation warnings when
running specs. It looks like a more involved upgrade of RSpec is
required to resolve these.