Fix some deprecation warnings
The recent Rails upgrade caused some deprecation warnings to appear.
This commit fixes some of them, making our RSpec runs quite a bit
cleaner. There are still some more that should be resolved, but I think
that will include a more involved upgrade of RSpec and the related RSpec
gems.

Flipping the switch to raise in transactional callbacks caused some
specs to fail because of file not found errors. I fixed this by moving
some stubbing up to happen for every spec. This threw off the keyboard
shortcuts feature spec which I fixed by adjusting it to avoid an
assertion that we don't actually care about.